% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_typeforms.R
\name{get_all_typeforms}
\alias{get_all_typeforms}
\title{Fetch all available typeforms}
\usage{
get_all_typeforms(api)
}
\arguments{
\item{api}{Your private api key.}
}
\value{
A two column data frame
}
\description{
This functions returns a data frame containing your typeforms and their
associated UIDs.
}
\examples{
\dontrun{
api = "XXXXX"
get_all_typeforms(api)
}
}

