% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Download results}
\usage{
get_results(uid, api, completed = NULL, since = NULL, until = NULL,
  offset = NULL, limit = NULL, order_by = NULL, simplify = TRUE)
}
\arguments{
\item{uid}{The UID (unique identifier) of the typeform you want the results for.}

\item{api}{Your private api key.}

\item{completed, }{default \code{NULL}, return all results.
Fetch only completed results (\code{TRUE}), or only not-completed results
(=\code{FALSE}).}

\item{since, }{default \code{NULL}}

\item{until, }{default \code{NULL}}

\item{offset}{Fetch all results except the first \code{offset}.
i.e. Start listing results from result #\code{offset} onwards.}

\item{limit, }{default \code{NULL}}

\item{order_by}{One of "completed", "date_land_desc", "date_land_incr",
"date_submit_desc", "date_submit_incr"}

\item{simplify, }{Logical. By default, \code{TRUE}, and returns only the questionnaire
responses as a data frame.}
}
\description{
Download results for a particular typeform.
}

